Reworked Algorithm to include multiple transports, various bugfixes
Hard to determine how well this is running at the moment because there's no graphical representation to eyeball and see any obvious errors. Found quite a few errors while updating the code so very likely there are still a few in there but the algorithm appears to be doing its job. The population tends towards a better weighted route. The breeding algorithm seems to correctly combine the mother/father genes and create viable and valid offspring. I will upload a published exe on discrod that should be easy to run. You can input 0 for default parameters or enter a population size, generations to run, number of points and number of transports.