Changes to UI, routing variables, and genetic route solver
Created 'special' passengers and transport agents. Special passengers can only be moved by special transport agents. Special is just a bool value.
Created some visual cues for what agents and destinations are special. (spherical agents and red cylinder destinations)
Created a new variables UI that lets you modify the constraints. number of points, number of agents, capacity of agents, number of special agents and points, and a reset to defaults. Modifying points necessarily sets randomise to on which uses the already created random point generation method.
Made some changes to MasterRoutingAgent that allows you to pass destinationmessages instead of points as a variable, this lets you account for special passengers/destinations in your routesolver.
New methods in the transportAgent and passengerAgent to set their special status and modify capacity based on PlayerPrefs set by the variable ui.
Deleted all the in scene agents, which are now generated when the scene loads. Passengers are automatically generated by stops and agents generated from PlayerPrefs.
Changes to the genetic solver that allow it to account for special transport and destinations.
At the moment the other routes do not use the special values, there is another PlayerPref that enables this functionality that can be added if the routesolver is able to account for them.
The K Means Cluster using the new genetic solver is now broken, but I will either add in non-special methods for the genetic solver or see if I can find a way to incorporate special agents into the Cluster solver. However the way it works might make this unfeasible.