Merge branch 'save-load' into Merge-UI-Additions-Save-Load
Had to deal with some conflicts.
Added special variables to the save-load.
Added a Mode variable that can be "Load" or "Generate" and changed MasterRoutingAgent/TransportNetwork to deal with either case.
Changed agents so they can be assigned values from PlayerPrefs or loaded from file.
Added a ui that displays the current variables during routing, number of agents, capacity etc.