Reverse the load order of items from file
file was written and read like a stack so each time you save/load it would reverse the order of passengers and transport agents which was causing trouble with the genetic algorithm. Changed it so when loading from file it reads from the bottom up so order is preserved between saving and loading.